server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
boto3==1.28.64
openai==1.3.0
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.3
httpx==0.25.1
